Fix another warning when comparing size_t and qsizetype
Amends 07f7ed2badf0cc1972bf6ba15f4c0cde4a773f19.

+++ tests/auto/tools/moc/tst_moc.cpp
@@ -2504,7 +2504,7 @@ void tst_Moc::cxx11Enums()
         QCOMPARE(size_t(metaType.sizeOf()), sizeof(char));
         QCOMPARE(isUnsigned, !std::is_signed_v<char>);
     } else if (isScoped) {
-        QCOMPARE(metaType.sizeOf(), sizeof(int));
+        QCOMPARE(size_t(metaType.sizeOf()), sizeof(int));
         QCOMPARE(isUnsigned, !std::is_signed_v<int>);
     } else {
         // underlying type is implementation defined
